+++ packages/measure/src/measure-function.tsx
@@ -6,7 +6,7 @@ import { showFlagsOutputIfNeeded, writeTestStats } from './output';
 export interface MeasureFunctionOptions {
   runs?: number;
   warmupRuns?: number;
-  dropOutliers?: boolean;
+  removeOutliers?: boolean;
   writeFile?: boolean;
 }
 
@@ -23,7 +23,7 @@ export async function measureFunction(fn: () => void, options?: MeasureFunctionO
 function measureFunctionInternal(fn: () => void, options?: MeasureFunctionOptions): MeasureResults {
   const runs = options?.runs ?? config.runs;
   const warmupRuns = options?.warmupRuns ?? config.warmupRuns;
-  const dropOutliers = options?.dropOutliers ?? config.dropOutliers;
+  const removeOutliers = options?.removeOutliers ?? config.removeOutliers;
 
   showFlagsOutputIfNeeded();
 
@@ -37,5 +37,5 @@ function measureFunctionInternal(fn: () => void, options?: MeasureFunctionOption
     runResults.push({ duration, count: 1 });
   }
 
-  return processRunResults(runResults, { warmupRuns, dropOutliers });
+  return processRunResults(runResults, { warmupRuns, removeOutliers });
 }